Revert handleFallDamage param to float — Mixin runtime confirms 1.21.4 method is still (FFLnet/minecraft/class_1282;)
The launcher crash log showed:
  Expected (FF...)  ← real method signature
  Found    (DF...)  ← my mixin

So in 1.21.4 yarn it's still (float fallDistance, float damageMultiplier, DamageSource source),
not (double, float, source) — that change must be in 1.21.5+. Reverted to float.

PlayerEntity check is preserved (so the integrated server's ServerPlayerEntity is also
covered in singleplayer/LAN — that part of the previous commit was correct).

+++ src/main/java/com/nxclient/mixin/LivingEntityMixin.java
@@ -13,7 +13,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class LivingEntityMixin {
 
     @Inject(method = "handleFallDamage", at = @At("HEAD"), cancellable = true)
-    private void onFallDamage(double fallDistance, float damageMultiplier, DamageSource source,
+    private void onFallDamage(float fallDistance, float damageMultiplier, DamageSource source,
                               CallbackInfoReturnable<Boolean> cir) {
         if (!NoFall.active) return;
         LivingEntity self = (LivingEntity) (Object) this;
